+++ lib/build_eval/server/team_city.rb
@@ -4,15 +4,12 @@ module BuildEval
     class TeamCity
 
       def initialize(args)
+        @http     = BuildEval::Http.new(args)
         @base_uri = args[:uri]
-        @username = args[:username]
-        @password = args[:password]
       end
 
       def build_result(name)
-        response = BuildEval::Http.get(
-          "#{@base_uri}/httpAuth/app/rest/buildTypes/id:#{name}/builds", username: @username, password: @password
-        )
+        response = @http.get("#{@base_uri}/httpAuth/app/rest/buildTypes/id:#{name}/builds")
         build_element = Nokogiri::XML(response.body).xpath("//build").first
         raise "Unexpected build response: #{response.message}" unless build_element
         BuildEval::Result::BuildResult.create(build_name: name, status_name: build_element.attribute("status").value)