+++ spec/lib/build_eval/http_shared_context.rb
@@ -1,0 +1,7 @@
+shared_context "stubbed http interactions" do
+
+  let(:http) { instance_double(BuildEval::Http) }
+
+  before(:example) { allow(BuildEval::Http).to receive(:new).and_return(http) }
+
+end