+++ spec/spec_helper.rb
@@ -15,3 +15,7 @@ if ENV["coverage"]
 end
 
 require_relative '../lib/build_eval'
+
+%w{ shared_examples shared_context }.each do |file_type|
+  Dir[::File.expand_path("../**/*_#{file_type}.rb", __FILE__)].each { |file| require file }
+end